+++ packages/designer/src/components/regions/workspace.vue
@@ -3,107 +3,52 @@
     class="v-workspace-region"
     :items="tabs"
     :menus="menus"
-    v-model="currentTab"
+    v-model="activeFileId"
     checkable
-    @remove="onRemove"
-    @command="onCommand"
+    @remove="onCloseTab"
     @action-click="onActionClick">
+    <template #menus>
+      <XAction
+        v-for="item of menus"
+        mode="icon"
+        size="small"
+        :type="menuChecked === item.name ? 'success' : 'info'"
+        background="always"
+        :label="item.label"
+        @click="onMenuChecked(item)">
+      </XAction>
+    </template>
     <template v-for="widget in widgets" :key="widget.name">
       <WidgetWrapper
         ref="widgetsRef"
-        v-if="currentTab === widget.name"
-        :key="computedDesignerKey"
+        v-if="refresh && menuChecked === widget.name"
+        :key="`${menuChecked}_${activeFileId}`"
         :region="region"
-        :widget="{ ...widget, props: openWidgetProps }"></WidgetWrapper>
+        :widget="{ ...widget, props: {} }"></WidgetWrapper>
     </template>
   </Tabs>
 </template>
 <script lang="ts" setup>
-  import { computed, reactive, ref } from 'vue';
-
+  import { ref, nextTick } from 'vue';
+  import { XAction } from '@vtj/ui';
   import { Tabs } from '../shared';
-  import { RegionType, type TabWidget } from '../../framework';
+  import { RegionType } from '../../framework';
   import { WidgetWrapper } from '../../wrappers';
-  import { useRegion } from '../hooks';
+  import { useRegion, useWorkspace } from '../hooks';
+
   export interface Props {
     region: RegionType;
   }
 
   const props = defineProps<Props>();
-
   const { widgets, widgetsRef } = useRegion(props.region);
-
-  const designerKey = ref(Symbol());
-
-  const items = reactive(
-    (widgets.value as TabWidget[]).map((n) => {
-      return {
-        name: n.name,
-        command: n.name,
-        label: n.label,
-        closable: !!n.closable,
-        props: n.props || {},
-        checked: !n.closable,
-        actions: n.actions as any[]
-      };
-    })
-  );
-
-  const tabs = computed(() => {
-    return items?.filter((n) => !n.closable || n.checked);
-  });
-
-  const menus = computed(() => {
-    return items.map((n) => {
-      return {
-        ...n,
-        disabled: !n.closable
-      };
-    });
-  });
-
-  const currentTab = ref(tabs.value[0]?.name);
-
-  const computedDesignerKey = computed(() => {
-    if (currentTab.value === 'Designer') {
-      return designerKey.value;
-    }
-    return;
-  });
-
-  const openWidgetProps = computed(() => {
-    const item = items.find((n) => n.name === currentTab.value);
-    return item?.props || {};
-  });
-
-  const onRemove = (name: any) => {
-    const item = items.find((n) => n.name === name);
-    if (item) {
-      item.checked = false;
-    }
-    if (currentTab.value === name) {
-      currentTab.value = tabs.value[0]?.name;
-    }
-  };
-  const onCommand = (e: any) => {
-    const item = items.find((n) => n.name === e.name);
-    if (item && item.closable) {
-      item.checked = !item.checked;
-      if (item.checked) {
-        currentTab.value = item.name;
-      }
-      if (!item.checked && currentTab.value === item.name) {
-        currentTab.value = tabs.value[0]?.name;
-      }
-    }
-  };
-
-  const openTab = (name: string, props: Record<string, any> = {}) => {
-    const item = items.find((n) => n.name === name);
-    if (item) {
-      item.props = Object.assign({}, item.props, props);
-      item.checked = true;
-      currentTab.value = item.name;
+  const { tabs, menus, activeFileId, menuChecked, onCloseTab, onMenuChecked } =
+    useWorkspace(widgets);
+  const refresh = ref(true);
+  const openTab = (_name: string, props: Record<string, any> = {}) => {
+    const url = props.url;
+    if (url) {
+      window.open(url);
     }
   };
 
@@ -119,14 +64,14 @@
     }
   };
 
-  const reload = () => {
-    if (currentTab.value === 'Designer') {
-      designerKey.value = Symbol();
-    }
+  const reload = async () => {
+    refresh.value = false;
+    await nextTick();
+    refresh.value = true;
   };
 
   const isDesignerActive = () => {
-    return currentTab.value === 'Designer';
+    return menuChecked.value === 'Designer';
   };
 
   defineOptions({
@@ -134,7 +79,7 @@
   });
 
   defineExpose({
-    currentTab,
+    currentTab: activeFileId,
     widgets,
     widgetsRef,
     openTab,