+++ packages/local/src/launch/index.ts
@@ -1,0 +1,3 @@
+export * from './web';
+export * from './h5';
+export * from './uniapp';