+++ packages/renderer/src/plugins/access.ts
@@ -161,7 +161,7 @@ export class Access {
   connect(params: AccessConnectParams) {
     const { mode, router, request } = params;
     this.mode = mode;
-    if (router && this.mode === ContextMode.Raw) {
+    if (router && this.mode !== ContextMode.Design) {
       this.setGuard(router);
     }
     if (request) {
@@ -313,6 +313,9 @@ export class Access {
       const id = to.params.id;
       return id && this.can(id);
     }
+    if (to.meta.__vtj__) {
+      return this.can(to.meta.__vtj__ as string);
+    }
     return true;
   }
 
@@ -324,13 +327,13 @@ export class Access {
     if (this.isWhiteList(to) || this.isAuthPath(to)) {
       return next();
     }
-
     if (this.isLogined()) {
       if (this.hasRoutePermission(to)) {
         return next();
       } else {
         const { noPermissionMessage = '无权限访问', unauthorized = false } =
           this.options;
+
         await this.showTip(noPermissionMessage);
         if (isFunction(unauthorized)) {
           unauthorized();
@@ -343,6 +346,7 @@ export class Access {
       }
     }
     next(false);
+    await delay(0);
     this.toLogin();
   }
 
@@ -381,6 +385,8 @@ export class Access {
         title: '提示',
         type: 'warning'
       })?.catch(() => false);
+    } else {
+      window.alert(content);
     }
     return false;
   }