+++ dev/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.5](https://github.com/samchen08/vtj.pro/compare/dev-web@0.13.4...dev-web@0.13.5) (2025-08-07)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.13.4](https://github.com/samchen08/vtj.pro/compare/dev-web@0.13.3...dev-web@0.13.4) (2025-08-07)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.13.3](https://github.com/samchen08/vtj.pro/compare/dev-web@0.13.2...dev-web@0.13.3) (2025-08-07)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.13.2](https://github.com/samchen08/vtj.pro/compare/dev-web@0.13.1...dev-web@0.13.2) (2025-08-06)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.13.1](https://github.com/samchen08/vtj.pro/compare/dev-web@0.13.0...dev-web@0.13.1) (2025-08-06)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+# [0.13.0](https://github.com/samchen08/vtj.pro/compare/dev-web@0.12.70...dev-web@0.13.0) (2025-08-06)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
 ## [0.12.70](https://github.com/samchen08/vtj.pro/compare/dev-web@0.12.69...dev-web@0.12.70) (2025-08-01)
 
 **Note:** Version bump only for package dev-web