+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.5](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.13.4...@vtj/icons@0.13.5) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/icons
+
+
+
+
+
+## [0.13.4](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.13.3...@vtj/icons@0.13.4) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/icons
+
+
+
+
+
+## [0.13.3](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.13.2...@vtj/icons@0.13.3) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/icons
+
+
+
+
+
+## [0.13.2](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.13.1...@vtj/icons@0.13.2) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/icons
+
+
+
+
+
+## [0.13.1](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.13.0...@vtj/icons@0.13.1) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/icons
+
+
+
+
+
+# [0.13.0](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.12.70...@vtj/icons@0.13.0) (2025-08-06)
+
+
+### Features
+
+* ✨ add pinia依/更新图标 ([d9b053e](https://gitee.com/newgateway/vtj/commits/d9b053ebcffbc457298a391d41cc01c03abeeaf5))
+
+
+
+
+
 ## [0.12.70](https://gitee.com/newgateway/vtj/compare/@vtj/icons@0.12.69...@vtj/icons@0.12.70) (2025-08-01)
 
 **Note:** Version bump only for package @vtj/icons