+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.5](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.13.4...@vtj/parser@0.13.5) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+## [0.13.4](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.13.3...@vtj/parser@0.13.4) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+## [0.13.3](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.13.2...@vtj/parser@0.13.3) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+## [0.13.2](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.13.1...@vtj/parser@0.13.2) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+## [0.13.1](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.13.0...@vtj/parser@0.13.1) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+# [0.13.0](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.70...@vtj/parser@0.13.0) (2025-08-06)
+
+
+### Features
+
+* ✨ project支持activeFiles ([3bb2b02](https://gitee.com/newgateway/vtj/commits/3bb2b0243dd601f15fb71d9b17f7883cf5511a4f))
+
+
+
+
+
 ## [0.12.70](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.69...@vtj/parser@0.12.70) (2025-08-01)
 
 **Note:** Version bump only for package @vtj/parser