+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.5](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.13.4...@vtj/ui@0.13.5) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.13.4](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.13.3...@vtj/ui@0.13.4) (2025-08-07)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.13.3](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.13.2...@vtj/ui@0.13.3) (2025-08-07)
+
+
+### Bug Fixes
+
+* 🐛 mask pure 改为路由meta控制 ([aebcc7b](https://gitee.com/newgateway/vtj/commits/aebcc7b4f0021bc2ae6147007568d6f4417a94a6))
+* 🐛 tab style ([b76a97c](https://gitee.com/newgateway/vtj/commits/b76a97c4ab500aad6984ae74c702292066f66c79))
+
+
+
+
+
+## [0.13.2](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.13.1...@vtj/ui@0.13.2) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.13.1](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.13.0...@vtj/ui@0.13.1) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+# [0.13.0](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.12.70...@vtj/ui@0.13.0) (2025-08-06)
+
+
+### Bug Fixes
+
+* 🐛 initRuntimeGlobals ([c8cfc9b](https://gitee.com/newgateway/vtj/commits/c8cfc9bae6cef714ede1e737025dc98a98808b9e))
+
+
+### Features
+
+* ✨ project支持activeFiles ([3bb2b02](https://gitee.com/newgateway/vtj/commits/3bb2b0243dd601f15fb71d9b17f7883cf5511a4f))
+
+
+
+
+
 ## [0.12.70](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.12.69...@vtj/ui@0.12.70) (2025-08-01)
 
 **Note:** Version bump only for package @vtj/ui